Robust test env setup
The test environment needs an isolated NODENV_ROOT, an unaltered
NODENV_VERSION, nodenv to be in PATH (from npm), for the plugin under
test to be in PATH, and for the plugin under test's hooks to be in
NODENV_HOOK_PATH.

An alternative would be to just symlink the plugin under test into a
dummy NODENV_ROOT/plugins. However, that's much slower than just setting
the env vars. Further, the explicit variables more clearly document
what's necessary for a plugin integration.