Shortcut to speed up lookup for common scenario
Most #engine version specs just specify a baseline version,
which means most likely, the highest installed version will satisfy
This does a first pass with just that single version in hopes it satisfies.
If so, we can avoid the cost of sh-semver sorting and validating across
all the installed versions.