Rewrite version-name hook to use new package command
Separately; old code had a bug where the first conditional branch would
never be executed (the `get_version_respecting_precedence` function
never returned non-zero). This is actually a good thing, because we
wouldn't _want_ to print anything to STDERR in this hook, anyway.

Also, this hook script is sourced by nodenv so it shouldn't have a
shebang, nor be executable.

Lastly, we rely on nodenv ensuring each plugin's `bin/` dir is added to
PATH so we can just invoke `nodenv-package` directly.