Explicit bin
npm won't include any symlinks in the tarball, so we need to explicitly
include the main bin. Pointing to libexec allows npm to make the symlink
from node_modules/.bin/ -> libexec/

The other bins aren't necessary because they are provided by npm
automaticaly (and included in path that's to npm). (Assuming this is
installed globally, which would ensure the bins are in PATH.)