Make version-origin output more consistent with nodenv itself
nodenv-version does not print the contents of .node-version nor
NODENV_VERSION files/variables. Nor does it indicate which
command/plugin is setting the value.

Instead, it prints the path to the file or the env var name. Users can
then inspect the file/variable themselves.

Similarly, we now print the path to the exact package.json file used,
and the #engines.node property that sets the value. We no longer print the
value of that property.